[Fix] Move RegisterParsedOpMetaCache to custom_operator.cc
Fix CI Linux-CPU linker error: undefined reference to
RegisterParsedOpMetaCache. The function was defined in Python binding
(eager_functions.cc) but called from inference library (custom_operator.cc).
libpaddle_inference.so doesn't link against Python bindings, causing
the linker error.

Solution: Move the function implementation and related types to
custom_operator.cc/h where they are accessible to both components.
